chore: More tests for add_identifier_data
